Add Indigo CDP and StabilityPool (#13)
* Add Indigo CDP and StabilityPool

They use generic Map which we don't support. We need first find example datum, so we are sure it's actually used so we can implement it correctly.

When inserting iAsset to a stability pool, datum of type Account is created

* Generic array

* Generic map

* Fix when run from console

cbor.entries.map is not a function or its return value is not iterable

From IDE it worked fine even before - of course Iterator has map function